Random functions
Description: The package random being imported comes with a lot of functions to be called. 'guess is a function and x is the parameter that a number will be passed to. random.randint(a, b) returns a random integer N such that a <= N <= b. a and b are the random number in the argument. The function will return a random number for the user to guess.