while loop
Description: Second step is once the computer has a random number the user needs to guess in terminal and user needs to input what that guess is and the computer will respond and tell user if it is too high or too low or the correct number. The program needs to loop until the user guess the right number. While loop is used because there is no predefined number to iterate over. Expression after the while keyword is written in a way so it will continue to loop as long as the random number is not guessed. guess 0 cause we don't want our guess to ever equal 0.